chat_server/pyproject.toml
[tool.poetry]
name = "django-chat"
version = "0.1.0"
description = ""
authors = ["genagurbanguliyev <genagurbanguliyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
channels = "^4.1.0"
channels-redis = "^4.2.0"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
django-environ = "^0.11.2"
django = "4.2"
psycopg = "^3.1.18"
daphne = "^4.1.2"
django-redis = "^5.4.0"
django-shortuuidfield = "^0.1.3"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
